requirements.txt
airia-sdk>=0.1.0
mcp>=0.1.0
aiohttp>=3.9.0
pydantic>=2.0.0
python-dotenv>=1.0.0
asyncio>=3.4.3
typing-extensions>=4.8.0